Make `load_payload()` signature consistent between types (#75)
Fixes #74.

The `JSONWebSignatureSerializer.load_payload` signature change is potentially breaking if anybody passed `return_header` as a second non-keyword argument to the method in the past. `itsdangerous.py` itself is using kwargs correctly everywhere so this shouldn't be an issue.

For the mixin, the change ensures it will pass additional arguments correctly.