change default intermediate hash from SHA-1 to SHA-512
Note that HMAC is still secure even with MD5. This is mostly a change
for aesthetics so people will stop asking, although it doesn't hurt to
be preemtive in this case.

Skipping SHA-256 for SHA-512 since it's more efficient on 64-bit.

See https://mail.python.org/pipermail/cryptography-dev/2017-March/000737.html for discussion.