Add the ability to specify the timestamp used when signing
This makes it much easier to test, as it is now possible to create
signed tokens that have a specific age. It also allows for tokens to
be created for a specific expiry time in mind, for example by creating
a token that has a midnight timestamp and a max age of 24 hours.